Docblockr: load parsers from file and store them in a Proxy
As the folder structure and filenames were changed, this might look like a bigger change than it really is.
Only `Docblockr` and the test files hold any line changes.

This makes it easier to add support for new Grammars.
Also adds a few new features:
- `require` and creation of parser instances are defered as much as possible
- instances are cached and no longer recreated when needed
- if Grammar is unsupported, the user is notified and `simple_mode` is enabled in that scope.

Moving `simple_mode` checks out of `docsparser` will be done in another commit.

About `grammars.json`:
It holds an array, in which each element represents a supported Grammar as an array of names.
The first value is the parser files and Grammars name.
Rest of the values are names of other Grammars, that should be aliased to / handled by the same parser.