docblockr-worker: convert `getParser` into a getter
As the folder structure and filenames were changed, the commit might look bigger than it really is.
Only `docblockr-worker` and `docsparser` hold any relevant changes.

The folder and parser names we're changed to more closely match the _Grammars_ they parse.

This makes it easier to add support for new Grammars.
Also adds a few new features:
- `require`ing of parsers is deferred as much as possible
- if the active Grammar is unsupported, the user is notified and `simple_mode` is enabled for it.